hw 3.4

+++ src/main/java/ru/hogwarts/school/service/FacultyService.java
@@ -2,7 +2,9 @@ package ru.hogwarts.school.service;
 
 import org.springframework.stereotype.Service;
 import ru.hogwarts.school.model.Faculty;
+import ru.hogwarts.school.model.Student;
 import ru.hogwarts.school.repository.FacultyRepository;
+import ru.hogwarts.school.repository.StudentRepository;
 
 import java.util.*;
 
@@ -10,9 +12,11 @@ import java.util.*;
 public class FacultyService {
 
     private final FacultyRepository facultyRepository;
+    private final StudentRepository studentRepository;
 
-    public FacultyService(FacultyRepository facultyRepository) {
+    public FacultyService(FacultyRepository facultyRepository, StudentRepository studentRepository) {
         this.facultyRepository = facultyRepository;
+        this.studentRepository = studentRepository;
     }
 
     public Faculty addFaculty(Faculty faculty) {
@@ -20,7 +24,7 @@ public class FacultyService {
         return facultyRepository.save(faculty);
     }
 
-    public Optional<Faculty> getFacultyInfo(long id) {
+    public Optional<Faculty> getFaculty(long id) {
         return facultyRepository.findById(id);
     }
 
@@ -33,7 +37,9 @@ public class FacultyService {
 
     public Optional<Faculty> removeFaculty(long id) {
         Optional<Faculty> deletedFaculty = facultyRepository.findById(id);
-        facultyRepository.deleteById(id);
+        if (deletedFaculty.isPresent()) {
+            facultyRepository.deleteById(id);
+        }
         return deletedFaculty;
     }
 
@@ -41,9 +47,15 @@ public class FacultyService {
         return facultyRepository.findAll();
     }
 
-    public Set<Faculty> getFaculties(String nameOrColor) {
-        Set<Faculty> set = new HashSet<>(facultyRepository.findByNameIgnoreCase(nameOrColor));
-        set.addAll(facultyRepository.findByColorIgnoreCase(nameOrColor));
-        return set;
+    public Collection<Faculty> getFacultiesByColor(String color) {
+        return facultyRepository.findFacultiesByColorIgnoreCase(color);
+    }
+
+    public Collection<Faculty> getFacultiesByNameOrColor(String nameOrColor) {
+        return facultyRepository.findFacultiesByNameIgnoreCaseOrColorIgnoreCase(nameOrColor, nameOrColor);
+    }
+
+    public Collection<Student> getStudents(Faculty faculty) {
+        return studentRepository.findStudentsByFaculty(faculty);
     }
 }
